instances: Expander.UnknownModuleInstances
This new method effectively returns a set of all of the instances of a
particular module whose addresses can't be fully determined yet, but does
so in the form of a set of patterns that match all of those as-yet-unknown
instance addresses.

So far this isn't used anywhere but in a future commit Terraform Core can
use this to generate placeholder unknown values when walking inside
unexpanded modules once we make it legal to have an unknown value for
the repetition argument of a module call.