instances: RepetitionData placeholders to use for unknown expansion
Now that Expander can support unknown expansions for modules and resource
instances, we'll need some placeholders to use when evaluating the
configurations of those objects to predict as much as possible what the
final configurations of all of their instances will look like.

This is a similar problem to checking whether a not-yet-expanded object
has valid configuration during the "validate" phase: we're essentially
asking what configuration all instances of the object will have in common,
and leaving unknown anything that might vary between instances.

Nothing is using these yet, but uses will follow in later commits.