core: Connect resource instance nodes with their expanders during apply
This was inadvertently trying to compare the string representation of an
addrs.AbsResource with the string representation of an
addrs.ConfigResource, making the two not match whenever the containing
module has count or for_each set.

The effects of this were pretty subtle:
 - During normal apply the whole-resource state updates were not
   necessarily happening before the instances got evaluated, but we got
   away with it most of the time because the provider configuration for
   a resource rarely changes and because we were redundantly
   re-calculating the for_each/count for every instance and thus not
   relying on the expander's idea of the instance expansion.
 - During destroy-time apply the pruning transformer was pruning the
   expansion nodes out altogether due to them having no dependents, and
   so we were not updating the state and expander at all in that case.

This now correctly ensures that the expander node runs first and so gets
a chance to update the resource's selected provider configuration and the
expander's sense of the resource expansion before we visit any
instance-specific code that relies on those.

This also switches to using `addrs.Map` instead of a direct Go map because
that allows the Go compiler to to type-check that we're using the address
types consistently, thereby reducing the risk of regressing this again
in the future.

There are unfortunately no now unit tests here because as usual our graph
transformation process resists unit testing as a result of all of the
transformers being subtly interdependent. However, this didn't regress
any existing tests and I manually inspected the graphs of the
configurations that caused me to spot this in the first place to verify
that they are now being constructed correctly.