namedvals: A new package to encapsulate the tracking of named values
Currently we have a slightly different treatment for each of the different
kinds of named values, which makes it hard to generalize the
implementations of their very similar behaviors.

This package should over time take ownership of the problem of tracking
all three kinds of named value during a graph walk, with result data
structures that use this information (such as state and plan objects)
copying the relevant values out of here only once the graph walk is
complete.

Nothing is using this yet, though. Some retrofitting of this into
Terraform Core will follow in subsequent commits.