addrs: Getting the parent of a PartialExpandedModule
Traversing upward from a PartialExpandedModule is trickier than traversing
down because we need to deal with what happens if the traversal crosses
over the boundary from partial-expanded into fully-expanded.

To deal with that we end up having two different methods to handle the
two situations, and a third method to indicate which one to call.
Thankfully the need to ask for the parent of a partial-expanded module is
relatively rare -- mainly just for input variables whose definitions need
to eval in the parent module's scope -- so this awkward API shouldn't be
needed in two many places.