terraform: Initial implementation of partial-expanded input variables
This is mainly just a proof-of-concept of what it might look like to
generate graph nodes representing placeholders for objects in
not-fully-expanded modules. These new codepaths are not really accessible
yet because it's still invalid to have a module whose expansion is
unknown; we'll continue down this path further in later commits once
there's actually somewhere to save the partially-evaluated placeholder
values.