core: Start of integrating "partial eval" mode into the evaluator
This is a new mode for the evaluator where instead of returning
information about exact objects it'll return placeholder values that
represent potentially many different hypothetical objects all declared
from the same static configuration object, in situations where we don't
yet have enough information to expand all of the modules and their
contents.

So far only the GetInputVariable function actually knows how to deal with
this, so this is far from sufficient but is a reasonable starting point
just to establish that it's possible to get Terraform into this evaluation
mode when working with graph nodes that represent such placeholder objects.