core: Expression evaluator can handle partial-eval in GetModule
This is a totally different approach to GetModule which uses the
configuration and previously-registered expansion to determine what ought
to exist in our named values state, rather than treating the values in
the named values state as the source of truth.

As a result we get an overall simpler implementation which is able to
panic if other components aren't behaving correctly, and we can return
placeholder results in partial evaluation mode, at least as long as we're
working with a single-instance module.

There are some further opportunities for simplification and improving the
detail of the unknown results if we make broader changes in future, but
for the moment this is just enough to mimic the previous behavior using
a new strategy.